category-engine/requirements.txt
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
async-timeout==5.0.1
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
datasets==3.2.0
Deprecated==1.2.15
dill==0.3.8
filelock==3.16.1
Flask==3.1.0
frozenlist==1.5.0
fsspec==2024.9.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-pubsub==2.27.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.12.7
grpcio==1.68.1
grpcio-status==1.62.3
huggingface-hub==0.27.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
libcst==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
numpy==2.0.2
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.1.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.4.5
six==1.17.0
sympy==1.13.1
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
